pyproject.toml
[project]
name = "financial_tone"
version = "0.0.0"
description = "A validator to check LLM-generated text for a particular tone."
authors = [
    {name = "Cartesia AI", email = "kg@cartesia.ai"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.4.0",
    "transformers",
    "torch",
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]